refactor: split utils.ts god module into 12 domain modules (resolves #53)
609-line utils.ts split into 12 sub-modules:

- isMac.ts — isMac()
- utils.ts — sleep, base64Decode, zdecode, zencode, randomInt
- platformReader.ts — readLockfile, readConfigs, readLog, parseShardFromLogline, parseLockFile
- assetLookup.ts — getAgent, getRank, getMap, getSeasonDateById
- match.ts — extractPlayers, extractPlayerName, sortPlayersForProcessing
- matchStats.ts — calculateStatsForPlayer, getMatchResult, calculateStreak
- ranking.ts — calculateRanking
- partyDetection.ts — extractParties
- smurfDetection.ts — isSmurf
- agentStats.ts — getPlayerBestAgent, calculateBestAgents, calculateBestMaps
- store.ts — getStoreItemInfo
- penalties.ts — extractPenalties

Bugs fixed:
- Division by matches.length instead of validMatches in calculateStatsForPlayer
- Removed console.log in production (calculateRanking)
- Fixed non-null assertion on AFK player stats

Removed dead code: extractEncounters, calculateMostPlayedServer

Updated all callers to import from index.ts instead of utils.ts
