+++ src/utils/isMac.ts
@@ -1,0 +1,1 @@
+export const isMac = () => navigator.platform.toUpperCase().indexOf("MAC") >= 0;
