+++ src/utils/utils.ts
@@ -1,609 +1,27 @@
-import { localDataDir } from "@tauri-apps/api/path";
-import { readDir, readTextFile, readTextFileLines } from "@tauri-apps/plugin-fs";
+const sleep = (ms: number = 2000) => new Promise((resolve) => setTimeout(resolve, ms));
+
 import base64 from "base-64";
-import { Buffer } from "buffer";
-import pako from "pako";
 
-import lockfile from "../../lockfile.json";
-import configs from "../../tests/fixtures/local/configs.json";
-import ShooterGameLog from "../../tests/fixtures/ShooterGame.json";
-import agents from "../assets/agents.json";
-import maps from "../assets/maps.json";
-import ranks from "../assets/ranks.json";
-import seasons from "../assets/seasons.json";
-
-import type {
-	BestAgent,
-	BestMaps,
-	Encounters,
-	MatchResult,
-	MostPlayedServer,
-	Parties,
-	Penalties,
-	PlayerMatchStats,
-	PlayerRanking,
-	Rank,
-	Streak,
-} from "@/interface/utils.interface";
-import type {
-	Agent,
-	AgentStats,
-	CurrentGameMatchResponse,
-	CurrentPreGameMatchResponse,
-	Map,
-	MatchDetailsResponse,
-	PenaltiesResponse,
-	PlayerMMRResponse,
-	PlayerNamesReponse,
-	PlayerRow,
-	Result,
-	StorefrontResponse,
-} from "../interface";
-
-export const sleep = (ms: number = 2000) => new Promise((resolve) => setTimeout(resolve, ms));
-
-export const isMac = () => navigator.platform.toUpperCase().indexOf("MAC") >= 0;
-
-export const base64Decode = (input: string): string => {
+const base64Decode = (input: string): string => {
 	return base64.decode(input);
 };
 
-export const zdecode = (input: string): any => {
+const zdecode = async (input: string): Promise<any> => {
+	const { Buffer } = await import("buffer");
+	const pako = await import("pako");
 	return JSON.parse(pako.inflateRaw(Buffer.from(input, "base64"), { to: "string" }));
 };
 
-export const zencode = (input: any): string => {
+const zencode = async (input: any): Promise<string> => {
+	const { Buffer } = await import("buffer");
+	const pako = await import("pako");
 	return Buffer.from(pako.deflateRaw(Buffer.from(JSON.stringify(input), "utf-8"))).toString("base64");
 };
 
-export const randomInt = (min: number, max: number): number => {
+const randomInt = (min: number, max: number): number => {
 	const minCeiled = Math.ceil(min);
 	const maxFloored = Math.floor(max);
 	return Math.floor(Math.random() * (maxFloored - minCeiled) + minCeiled);
 };
 
-export const readLockfile = async (): Promise<string> => {
-	if (isMac()) {
-		return lockfile;
-	} else {
-		const path = await localDataDir();
-		const file = await readTextFile(path + "\\Riot Games\\Riot Client\\Config\\lockfile");
-		return file.toString();
-	}
-};
-
-export const readConfigs = async (): Promise<string[]> => {
-	if (isMac()) {
-		return configs;
-	} else {
-		const path = await localDataDir();
-		const files = await readDir(path + "\\Valorant\\Saved\\Config");
-		return files.map((file) => file.name).filter((name) => name.match(/(.*)-(.*)-(.*)-(.*)-(.*)/));
-	}
-};
-
-export const readLog = async () => {
-	if (isMac()) {
-		for (const line of ShooterGameLog) {
-			const res = parseShardFromLogline(line);
-			if (res) return res;
-		}
-	} else {
-		const path = await localDataDir();
-		const lines = await readTextFileLines(path + "\\Valorant\\Saved\\Logs\\ShooterGame.log");
-		for await (const line of lines) {
-			const res = parseShardFromLogline(line);
-			if (res) return res;
-		}
-	}
-
-	return ["", ""];
-};
-
-export const parseShardFromLogline = (line: string): [string, string] | undefined => {
-	if (!line.includes("https://glz")) return;
-
-	const urlMatch = line.match(/URL \[GET (https?:\/\/[^\]]+)\]/);
-	if (!urlMatch) return;
-
-	const regionMatch = urlMatch[1].match(/glz-([a-zA-Z]+)-\d+\.([a-zA-Z]+)\.a\.pvp\.net/);
-	if (regionMatch) return [regionMatch[1], regionMatch[2]];
-};
-
-export const parseLockFile = (content: string): { port: string; password: string } => {
-	const [_, __, port, password, ___] = content.split(":");
-
-	return { port, password: base64.encode(`riot:${password}`) };
-};
-
-export const extractPlayers = (
-	match: CurrentPreGameMatchResponse | CurrentGameMatchResponse | MatchDetailsResponse,
-): string[] => {
-	if ("AllyTeam" in match) return match.AllyTeam?.Players.map((player) => player.Subject) || [];
-
-	if ("Players" in match) return match.Players.map((player) => player.Subject);
-
-	return match.players.map((player) => player.subject);
-};
-
-export const calculateStatsForPlayer = (puuid: string, matches: MatchDetailsResponse[]): PlayerMatchStats => {
-	let validMatches = 0;
-	const stats = {
-		kills: 0,
-		deaths: 0,
-		assists: 0,
-		kd: 0,
-		hs: 0,
-		adr: 0,
-		wins: 0,
-		losses: 0,
-		ties: 0,
-		winrate: 0,
-	};
-
-	for (const match of matches) {
-		const player = match.players.find((player) => player.subject === puuid);
-
-		if (!player || !player.stats) continue;
-
-		validMatches += 1;
-		stats.kills += player.stats.kills;
-		stats.deaths += player.stats.deaths;
-		stats.assists += player.stats.assists;
-		stats.kd += player.stats.kills / (player.stats.deaths || 1);
-
-		const { result } = getMatchResult(puuid, match);
-		result == "won"
-			? (stats.wins += 1)
-			: result == "loss"
-				? (stats.losses += 1)
-				: result == "tie"
-					? (stats.ties += 1)
-					: null;
-
-		if (match.roundResults) {
-			const shots = { legshots: 0, bodyshots: 0, headshots: 0 };
-			let damagePerRound = 0;
-
-			for (const roundResult of match.roundResults) {
-				for (const damage of roundResult.playerStats.find((result) => result.subject === puuid)!.damage) {
-					shots.legshots += damage.legshots;
-					shots.bodyshots += damage.bodyshots;
-					shots.headshots += damage.headshots;
-					damagePerRound += damage.damage;
-				}
-			}
-
-			stats.adr += damagePerRound / match.roundResults.length;
-			stats.hs += shots.headshots / (shots.headshots + shots.bodyshots + shots.legshots);
-		}
-	}
-
-	if (validMatches == 0) return stats;
-
-	return {
-		kills: Math.round(stats.kills / matches.length),
-		deaths: Math.round(stats.deaths / matches.length),
-		assists: Math.round(stats.assists / matches.length),
-		kd: Number((stats.kd / matches.length).toFixed(2)),
-		hs: Math.round((stats.hs / matches.length) * 100),
-		adr: Math.round(stats.adr / matches.length),
-		wins: stats.wins,
-		losses: stats.losses,
-		ties: stats.ties,
-		winrate: Math.round((stats.wins / matches.length) * 100),
-	};
-};
-
-export const calculateRanking = (playerMMR: PlayerMMRResponse): PlayerRanking => {
-	const seasonalInfo = playerMMR?.QueueSkills.competitive.SeasonalInfoBySeasonID
-		? Object.values(playerMMR.QueueSkills.competitive.SeasonalInfoBySeasonID)
-		: [];
-
-	const getSeasonPeakTier = (season: (typeof seasonalInfo)[number]) =>
-		season.WinsByTier ? Math.max(...Object.keys(season.WinsByTier).map((tier) => Number(tier))) : season.CompetitiveTier;
-
-	const peakSeason = seasonalInfo.reduce<(typeof seasonalInfo)[number] | null>((best, season) => {
-		if (!best) return season;
-
-		const seasonPeakTier = getSeasonPeakTier(season);
-		const bestPeakTier = getSeasonPeakTier(best);
-
-		if (seasonPeakTier !== bestPeakTier) {
-			return seasonPeakTier > bestPeakTier ? season : best;
-		}
-
-		if (season.CompetitiveTier !== best.CompetitiveTier) {
-			return season.CompetitiveTier > best.CompetitiveTier ? season : best;
-		}
-
-		return season.RankedRating > best.RankedRating ? season : best;
-	}, null);
-
-	console.log(playerMMR)
-
-	return {
-		currentRank: playerMMR?.LatestCompetitiveUpdate?.TierAfterUpdate || 0,
-		currentRR: playerMMR?.LatestCompetitiveUpdate?.RankedRatingAfterUpdate || 0,
-		peakRank: peakSeason ? getSeasonPeakTier(peakSeason) : 0,
-		peakRankSeasonId: peakSeason?.SeasonID || null,
-		lastGameMMRDiff: playerMMR?.LatestCompetitiveUpdate?.RankedRatingEarned || 0,
-	};
-};
-
-export const getMatchResult = (puuid: string, match: MatchDetailsResponse): MatchResult => {
-	if (!match || !match.teams) return { result: "N/A", score: "", accountLevel: 0 };
-
-	const player = match.players.find((player) => player.subject === puuid)!;
-	const team = match.teams.find((team) => team.teamId === player.teamId)!;
-
-	if (match.teams[0].roundsWon === match.teams[1].roundsWon)
-		return {
-			result: "tie",
-			score: `${match.teams[0].roundsWon}:${match.teams[1].roundsWon}`,
-			accountLevel: player.accountLevel,
-		};
-
-	return {
-		result: team.won ? "won" : "loss",
-		score: `${team.roundsWon}:${team.roundsPlayed - team.roundsWon}`,
-		accountLevel: player.accountLevel,
-	};
-};
-
-export const getAgent = (uuid: string): Agent => {
-	if (!uuid) return { uuid: "", displayName: "", killfeedPortrait: "", displayIcon: "" };
-
-	return (
-		agents.find((agent) => agent.uuid === uuid.toLowerCase()) || {
-			uuid: "",
-			displayName: "New Agent",
-			killfeedPortrait: "",
-			displayIcon: "",
-		}
-	);
-};
-
-export const getRank = (rank: number): Rank => {
-	const _rank = ranks.find((_rank) => _rank.tier === rank);
-	return {
-		tier: rank,
-		rankName: _rank?.tierName as string,
-		rankColor: _rank?.color as string,
-		rankImg: _rank?.largeIcon as string,
-	};
-};
-
-export const getMap = (uuid: string): Map => {
-	const map = maps.find((map) => map.mapUrl === uuid);
-	return map || { uuid, displayName: "New Map", displayIcon: "", listViewIcon: "" };
-};
-
-export const getSeasonDateById = (seasonId: string): Date | null => {
-	const season = seasons.find((season) => season.seasonUuid === seasonId);
-	if (!season) return null;
-
-	return new Date(season.endTime);
-};
-
-export const isSmurf = (player: PlayerRow) => {
-	return Boolean(player.accountLevel && player.accountLevel < 100 && player.kd && player.kd > 1.5);
-};
-
-export const getPlayerBestAgent = (puuid: string, matches: MatchDetailsResponse[], mapUrl: string): AgentStats[] => {
-	matches = matches.filter((match) => match.matchInfo.mapId === mapUrl);
-
-	const agents: {
-		[key: string]: { k: number; d: number; kd: number; games: number };
-	} = {};
-
-	for (const match of matches) {
-		const player = match.players.find((player) => player.subject === puuid)!;
-
-		if (!player.stats) continue;
-
-		const { kills, deaths } = player?.stats;
-
-		if (!(player.characterId in agents)) {
-			agents[player.characterId] = { k: kills, d: deaths, kd: 0, games: 1 };
-			continue;
-		}
-
-		agents[player.characterId].k += kills;
-		agents[player.characterId].d += deaths;
-		agents[player.characterId].games += 1;
-	}
-
-	for (const characterId in agents) {
-		agents[characterId].k = Math.round(agents[characterId].k / agents[characterId].games);
-		agents[characterId].d = Math.round(agents[characterId].d / agents[characterId].games);
-		agents[characterId].kd = parseFloat((agents[characterId].k / agents[characterId].d).toFixed(2));
-	}
-
-	return Object.keys(agents)
-		.map((characterId) => ({
-			agentId: characterId,
-			agentUrl: getAgent(characterId).displayIcon!,
-			avgKills: agents[characterId].k,
-			avgDeaths: agents[characterId].d,
-			avgKd: agents[characterId].kd,
-			games: agents[characterId].games,
-		}))
-		.sort((a, b) => b.avgKd - a.avgKd);
-};
-
-export const calculateBestAgents = (puuid: string, matches: MatchDetailsResponse[]): BestAgent[] => {
-	const bestAgents: BestAgent[] = [];
-
-	const agents: { [key: string]: MatchDetailsResponse[] } = {};
-
-	for (const match of matches) {
-		const player = match.players.find((player) => player.subject === puuid)!;
-
-		if (!(player.characterId in agents)) {
-			agents[player.characterId] = [match];
-		} else {
-			agents[player.characterId].push(match);
-		}
-	}
-
-	for (const agentId in agents) {
-		const stats = calculateStatsForPlayer(puuid, agents[agentId]);
-		bestAgents.push({
-			agentId,
-			matches: agents[agentId].length,
-			...stats,
-		});
-	}
-
-	return bestAgents.sort((a, b) => (b.matches || 1) - (a.matches || 0));
-};
-
-export const calculateBestMaps = (puuid: string, matches: MatchDetailsResponse[]): BestMaps[] => {
-	const bestMaps: BestMaps[] = [];
-
-	const maps: { [key: string]: MatchDetailsResponse[] } = {};
-
-	for (const match of matches) {
-		if (!(match.matchInfo.mapId in maps)) {
-			maps[match.matchInfo.mapId] = [match];
-		} else {
-			maps[match.matchInfo.mapId].push(match);
-		}
-	}
-
-	for (const mapId in maps) {
-		const stats = calculateStatsForPlayer(puuid, maps[mapId]);
-		bestMaps.push({
-			mapId,
-			matches: maps[mapId].length,
-			...stats,
-		});
-	}
-
-	return bestMaps.sort((a, b) => (b.matches || 1) - (a.matches || 0));
-};
-
-export const getStoreItemInfo = (
-	offers:
-		| StorefrontResponse["AccessoryStore"]["AccessoryStoreOffers"]
-		| StorefrontResponse["SkinsPanelLayout"]["SingleItemStoreOffers"]
-		| StorefrontResponse["BonusStore"]["BonusStoreOffers"],
-): {
-	uuid: string;
-	price: number;
-	type: "weaponskin" | "spray" | "playercard" | "buddy";
-}[] => {
-	if (!offers) return [];
-
-	return offers.map((offer) => {
-		const rawOffer = "Offer" in offer ? offer.Offer : offer;
-
-		return {
-			uuid: rawOffer.Rewards[0]["ItemID"],
-			price: Object.values(rawOffer.Cost)[0],
-			type:
-				rawOffer.Rewards[0]["ItemTypeID"] === "e7c63390-eda7-46e0-bb7a-a6abdacd2433"
-					? "weaponskin"
-					: rawOffer.Rewards[0]["ItemTypeID"] === "d5f120f8-ff8c-4aac-92ea-f2b5acbe9475"
-						? "spray"
-						: rawOffer.Rewards[0]["ItemTypeID"] === "3f296c07-64c3-494c-923b-fe692a4fa1bd"
-							? "playercard"
-							: "buddy",
-		};
-	});
-};
-
-export const extractPenalties = (penalties: PenaltiesResponse): Penalties | null => {
-	if (!penalties.Infractions.length) return null;
-
-	penalties.Penalties = penalties.Penalties.filter((p) => p.RiotRestrictionEffect);
-
-	const penalty = penalties.Penalties.find(
-		(penalty) => penalty.RiotRestrictionEffect.RestrictionType === "PBE_LOGIN_TIME_BAN",
-	);
-
-	if (!penalty) return null;
-
-	return {
-		freeTimestamp: +new Date(penalty.Expiry),
-		type: penalties.Penalties.map((p) => p.RiotRestrictionEffect.RestrictionType),
-		reason: penalties.Penalties.map((p) => p.RiotRestrictionEffect.RestrictionReason).filter((p) => p.length),
-		matchId: penalty.IssuingMatchID,
-	};
-};
-
-export const extractPlayerName = (
-	puuid: string,
-	matches: MatchDetailsResponse[],
-): { name: string; tag: string } | null => {
-	for (const match of matches) {
-		const player = match.players.find((player) => player.subject === puuid);
-		if (!player) continue;
-
-		if (player.subject !== "" && player.tagLine !== "") return { name: player.gameName, tag: player.tagLine };
-	}
-	return null;
-};
-
-export const calculateMostPlayedServer = (matches: MatchDetailsResponse[]): MostPlayedServer => {
-	const servers: MostPlayedServer = {};
-
-	for (const match of matches) {
-		const serverName = match.matchInfo.gamePodId.split("-")[4];
-
-		if (!(serverName in servers)) servers[serverName] = 0;
-
-		servers[serverName] += 1;
-	}
-
-	return servers;
-};
-
-export const calculateStreak = (puuid: string, matches: MatchDetailsResponse[]): Streak | null => {
-	if (!matches.length) return null;
-
-	const results: Result[] = [];
-
-	for (const match of matches) {
-		const { result } = getMatchResult(puuid, match);
-		results.push(result);
-	}
-
-	const streak: Streak = {
-		type: results[0],
-		number: 0,
-	};
-
-	for (const result of results) {
-		if (result === streak.type) {
-			streak.number += 1;
-		} else {
-			break;
-		}
-	}
-
-	return streak;
-};
-
-export const extractEncounters = (
-	puuid: string,
-	matches: MatchDetailsResponse[],
-	players: PlayerNamesReponse[],
-): Encounters[] => {
-	const encounters: { [key: string]: number } = {};
-
-	for (const match of matches) {
-		for (const id of extractPlayers(match)) {
-			if (id === puuid || !players.map((p) => p.Subject).includes(id)) continue;
-
-			if (id in encounters) {
-				encounters[id] += 1;
-			} else {
-				encounters[id] = 1;
-			}
-		}
-	}
-
-	return Object.keys(encounters).map((puuid) => {
-		const player = players.find((p) => p.Subject === puuid)!;
-		return {
-			puuid,
-			name: player.GameName,
-			tag: player.TagLine,
-			number: encounters[puuid],
-		};
-	});
-};
-
-export const extractParties = (
-	data: { puuid: string, matches: MatchDetailsResponse[] }[]
-): Parties => {
-	const connections = new Map<string, Set<string>>();
-	const currentLobbyIds = new Set(data.map(({ puuid }) => puuid));
-
-	for (const { puuid, matches } of data) {
-		if (!connections.has(puuid)) {
-			connections.set(puuid, new Set<string>());
-		}
-
-		for (const match of matches) {
-			const player = match.players.find((p) => p.subject === puuid);
-			if (!player?.partyId) continue;
-
-			for (const teammate of match.players) {
-				if (teammate.subject === puuid) continue;
-				if (!currentLobbyIds.has(teammate.subject)) continue;
-				if (teammate.partyId !== player.partyId) continue;
-				if (player.teamId && teammate.teamId && player.teamId !== teammate.teamId) continue;
-
-				connections.get(puuid)?.add(teammate.subject);
-
-				if (!connections.has(teammate.subject)) {
-					connections.set(teammate.subject, new Set<string>());
-				}
-
-				connections.get(teammate.subject)?.add(puuid);
-			}
-		}
-	}
-
-	const visited = new Set<string>();
-	const parties: Parties = [];
-	let nextPartyId = 1;
-
-	for (const puuid of currentLobbyIds) {
-		if (visited.has(puuid)) continue;
-
-		const stack = [puuid];
-		const component = new Set<string>();
-
-		while (stack.length) {
-			const current = stack.pop()!;
-			if (visited.has(current)) continue;
-
-			visited.add(current);
-			component.add(current);
-
-			for (const neighbor of connections.get(current) || []) {
-				if (!visited.has(neighbor)) {
-					stack.push(neighbor);
-				}
-			}
-		}
-
-		if (component.size < 2) continue;
-
-		parties.push({
-			partyId: nextPartyId,
-			puuids: [...component],
-		});
-
-		nextPartyId += 1;
-	}
-
-	return parties;
-};
-
-export const sortPlayersForProcessing = (
-	players: PlayerNamesReponse[],
-	table: Record<string, PlayerRow>,
-): PlayerNamesReponse[] => {
-	const playersObj: Record<string, PlayerNamesReponse> = {};
-	for (const player of players) {
-		playersObj[player.Subject] = player;
-	}
-
-	const tableAsArray = Object.values(table).sort((a, b) => {
-		if (a.enemy !== b.enemy) {
-			return +a.enemy - +b.enemy;
-		}
-
-		if (a.accountLevel && b.accountLevel) return a.accountLevel - b.accountLevel;
-		if (a.accountLevel) return -1;
-		if (b.accountLevel) return 1;
-		return 0;
-	});
-
-	return tableAsArray.map((p) => playersObj[p.puuid]);
-};
+export { sleep, base64Decode, zdecode, zencode, randomInt };
